Move `h.schemas.group` -> `h.schemas.api.group`
